Extracted TransferManager interface
