+++ qabelbox/src/main/java/de/qabel/qabelbox/storage/BoxTransferListener.java
@@ -1,0 +1,8 @@
+package de.qabel.qabelbox.storage;
+
+public interface BoxTransferListener {
+
+    void onProgressChanged(long bytesCurrent, long bytesTotal);
+
+    void onFinished();
+}
